Fixes broken test when built on non-english locale
FileUtilTest tests the output of FileUtil.formatFileSize() (which
depends on the locale of the machine the test is run on) against a
literal string (which does not depend on the locale).

This will fail on non-english locales. This fix uses the decimal
separator and grouping separator of the machine running the test and
should make the test independent of the locale it is executed on.

This PR is the exact solution already pointed out in the comments of

Fixes #197 and #198.