Add two utility methods to copy data from WindowReaders.
To support random access interfaces for various readers, we have to
wrap a WindowReader from the existing IdentificationRequest, and
provide the data to the archive class using a random access interface
it supports.

Two methods are provided, to copy bytes from a WindowReader into either
a byte array (with offset and length), or into a ByteBuffer.