Remove dependency on aspect oriented code generation.
Aspects were being used to inject code into a single method.
The whole point of aspect oriented programming (AOP) is to inject
common code across lots of methods (e.g. to log what they do),
in order to simplify managing cross-cutting concerns.

There is literally no point in having this complexity when you
only inject the code into a single method in the application.
This is achieved in a simpler and more consistent way by just
injecting the dependency into the class that needs it,
as is done in the rest of DROID.