Avoid parsing the container signature file more than once.
The current design parses the signature file when instantiating
the OLE2 container handler, and also the ZIP container handler.

This change splits out a new bean, the ContainerSignatureFileReader,
which parses the signature file on demand and caches the parsed
definitions the first time the file is parsed.

The AbstractContainerIdentifier then uses this class to obtain the
definitions.  spring-results.xml is modified to use the new bean,
and some tests needed changing to set the ContainerSignatureFileReader
instead of a parser directly.