Add parameterized constructor and remove synchronization.
No controversy in adding a parameterized constructor.

Synchronization is removed when getting the temp dir location.
This is a simple strategy class that is instantiated with the location
of the temp dir location (or should be set shortly after construction if
using bean style).  There is no need for a synchronization point when
getting this location for every single container file request.