Inherit from JDBCBatchResultHandler to get all the database specific methods.
Although it makes little sense for the result handler to be involved in
setting up new databases and loading formats, this is how the design currently works.
So any other result handler dao class needs to support database initialisation and
format loading from the database, even if it ultimately doesn't write any results
out to the database.  Refactoring this looks really quite complex.