FIXES: Any ?? matcher should not be part of sequences.
?? symbols were being incorrectly included as part of the XML signature text.  They should instead be modelled
as min/max offsets between what precedes them and what follows them.  This fix removes the ANY (??) matcher as a valid
part of an anchoring sequence, as that was incorrect.

The fix is to ensure that the ANY type is not processed as an anchor type, and cannot be included in fragment text either.  Finally, sigtool correctly takes account of the anchor type (BOF/EOF) when serializing signatures.