Add GZIP container identification.
This adds a few constants in to allow DROID to read GZIP container
signatures.

Most of the work is done in the new GzIdentifier class.

I've added tests where I can. The main thing stopping me was that there
are no GZIP container signatures in the PRONOM signature files because
DROID doesn't support them yet.

To get around this, I've added tests into the droid-api module which
generates a container file and a signature file and generates the
relevant zip/ole2/gzip file and tests against that. It's verbose but it
won't break when PRONOM is updated.

I used some records in the tests which is a Java 21 feature so I had to
update some of the plugins which then meant I had to update some
dependencies so that covers the rest of the changes.