Do not increment loop counter.
fixes #1140

DROID is getting stuck in an infinite loop trying to match the right fragments for [this signature](https://github.com/nationalarchives/pronom-signatures/blob/develop/signatures/fmt/134.json#L132)

The steps are something like:

1. It finds matches for 7 of the 8 right fragments.
2. It adds these to a matched list
3. It doesn't find a match for the 8th fragment
4. It then searches each of the fragments in the matched list to either side of the previous matches.
5. Fragments 1-7 match and so stay in the matched list
6. The last matched fragment position is set to 7
7. The loop checking each fragment increments the fragment position to 8
8. It doesn't find a match for the 8th fragment
9. Repeat steps 3 - 9

The problem I think is step 6. It shouldn't be changing the loop counter.

The issue is that I'm not 100% sure if by doing this, we'll be skipping
some fragment checks which will cause a different format to not be
matched at some point in the future.
I don't think so though because the main loop still iterates through
every fragment.