Add ZIP file fallback.
There are files which aren't quite zip files which truevfs won't open.

This change tries to open the file using the truevfs classes as usual.
If there is an error doing this, it will try again with the Apache commons
zip library which is a lot less fussy about what it opens.

I've tried it against one test file I had and this is working.

I've added in a test as well which creates a zip file with one entry and
then overrides the file count byte with 0 instead of 1. This causes it
to fail on the initial read and passes when read by commons-compress.