Fix the JSON output.
As it was, it was only outputting the last file if there was more than
one in the profile.

The problem with it is that you need to have a [ and ] at the beginning
and end to make it valid json and the objects with commas in between
which is hard to incrementally like we're doing with the csv output.

So I've changed this so that the json array is kept in memory and then
written once everything has finished processing. This is working.

It's possible that someone will have such a huge number of files that
you run out of memory. If that happens, I'll think of something else.