Fix the JSON output.
As it was, it was only outputting the last file if there was more than one in the profile.

This was because we were creating the OutputJson object for each file we were writing which created the json array node as new every time so we only ever got the last item.

I've changed this so it's only created once.

This then caused the problem that every intermediate json object was being written out. I've fixed this by using the JsonGenerator which streams each object to the writer.

This led to another problem in that a json array needs [ and ] wrapping it. It's simple to add one at the beginning. Adding one at the end for the gui export is fine because we know when the export is complete but it's more tricky with the CLI because all of the nodes in the profile run in different threads.

I added a method in the ProfileRunCommand which writes ] to the end of either the PrintWriter or FileWriter if the output is set to JSON. It's not elegant but it does work.