Add glob matching and container file name matching.
This is trying to solve the issue here https://github.com/digital-preservation/pronom/issues/10

There are two changes.

One is to do glob matching on container file names. In both the
Ole2IdentifierEngine and the ZipIdentifierEngine, we at first try to get
the name from the container as is. If that doesn't work we do:

* Check if the string `{containerFileName}` is in the path. If it is,replace the placeholder with the container file name and match on that.
* Check if the path in the container file matches and files in the container using glob matching.

This is experimental and subject to change but as it shouldn't break any
existing functionality, I think this can go into main.