UI and CLI tests
The UI tests use AssertJ Swing to test a running version of the UI.

The tests are executed inside a docker container which allows them to
run on GitHub runners where there is no UI set up.

The CLI tests use bats and run the CLI through a terminal rather than
calling the classes like the unit tests do.

These are meant to be tests made before a dev release to make sure
nothing is seriously broken. They are *not* meant to be completely
comprehensive, nor are they a replacement for manual testing before a
full release.