Trim file path with spaces at the end for Windows.
Should fix #1130

If you import a profile with a file with a space at the end from an OS
which allows it into DROID on Windows, the export will fail.

The file url comes in with %20 at the end which fails when calling
Path.get on Windows because spaces are not allowed at the end.

This detects if there's %20 at the end and trims it if it is running on
Windows.

The export now works for this scenario although there may be other
issues we've not found yet.