Stop writing JSON array to CSV output
Fixes #1318

The start and end array tokens for the json output are being written
when the FormattedDataWriter is being initialised. We're not checking to
see if the JSON output option is set. I've added a check on the open and
close methods which has stopped this.

The reason the tests didn't catch this is because the [] is written when
close() is called and this is only called after the assertions in the
tests. I've updated the tests to call close after each one and they're
passing now where they weren't before.