Detect signature file changes.
Relates to https://github.com/digital-preservation/droid/issues/1274

When DROID installs a signature file, it creates a profile template.
This profile template is a derby database will all of the file formats
stored in it.
Once this template exists, DROID will re-use it for any new profiles.
This causes the problem where if you change the signature file, DROID
won't pick up the changes. You have to delete the ~/.droid6 folder to
force DROID to rebuild the profile template database.

Now, when a signature file is loaded, a property is put into
droid.properties with the last modified date of the sig file.
When DROID starts or when a new profile is created, it checks the
current last modified against the one in the properties file. If the sig
file has been updated, it returns a status of NO_TEMPLATE and rebuilds
the template database.

This now means that if you make a change to the signature file, this
will be picked up with any new profile, without having to restart DROID
or clean the ~/.droid6 folder.