Create 'clean' option for DROID.
This will clean the folders out of the ~/.droid6 directory.

This is often what is recommended when DROID misbehaves so having a
command to do this will be useful.

It will delete the following directories and files

container_sigs/
signature_files/
profiles/
profile_templates/
droid.properties
log4j2.properties
tmp/