Implement reloadWindowBytes
There was a bug where DROID was returning fmt/189 for files which should
have returned fmt/1827.

The bytes that are returned from createWindow in the WindowReaders are
stored in a SoftReference inside SoftWindow. The idea being that if the
system starts to run out of memory, the byte arrays can be reclaimed by
the garbage collector.

If this happens, the byte array is null and getByteArray inside
SoftWindow calls reloadWindowBytes to repopulate the byte array.

This method was not implemented because I originally thought that it
wasn't used as it was never called when running the tests or during
manual testing. The reason it was never called is that my dev machine
has plenty of memory so the GC never had to reclaim anything.

When DROID was running in the lambda inside TDR, it only had 2Gb of
memory so for a larger file, it was running out of memory and the GC was
reclaiming the bytes. When it tried to call reloadWindowBytes to
repopulate it, it was returning an empty array and so DROID couldn't
identifiy the file.

I've not implemented this method. I've tested this on TDR and it's
returning the correct PUID.