Max bytes to scan changes
There are files going through TDR that are never returning a result
because the maxBytesToScan is set to Long.MaxValue. I think this might
be a signature problem but I'm not sure.

I've changed the default value to be the same as the DROID UI/CLI use
and also added in the ability to change it if necessary.

There's a couple of dependency updates which have got rid of the CVE
warnings as well.