Add debug mode
This was prompted by this issue https://github.com/digital-preservation/droid/issues/773

The way this works is:
You set a profile.debug property to true in droid.properties

This sets a spring profile called debug.

This loads the DebugFileSystemIdentificationRequest instead of the usual FileSystemIdentificationRequest

This delegates most methods to FileSystemIdentificationRequest. The only difference is that the WindowReader is now a DebugFileReader

The DebugFileReader again delegates most methods to FileReader. For each call to createWindow, it increments the bytes read. For each call
to getWindow, if the window is in the cache, it increments the cache bytes read.

Then as we loop over the signatures, if isDebug is true, we output the bytes read, cache bytes read and whether there was a match.

I've added the DebugFileSystemIdentificationRequest to the skeleton suite tests to make sure that it behaves the same.