Fix bug with DROID API.
We were calling prepareForUse() on the sig file class which adds the
file formats to a list for the signatures.

The problem was that the call to droidCore.init() already calls this.
This means that for every signature, we had the file format in there
twice. This then meant that no signatures with corresponding container
signatures were being removed from the file format list because there
was always at least one signature left.

This was mostly fine until now but the new signature file has a binary
signature for an Android app bundle which seems to match most containers
and so the API tests were failing. These tests are now passing.

Sadly, the bug meant that the API skeleton suite tests were passing
because none of the binary signatures were being removed for files with
container signatures. This is now not happening so I've added a big list
of exceptions to the tests.