Try to get the latest version from the json API.
This is in preparation for the new PRONOM service.

There is a json file at `/signatures.json` on the new service which
shows the versions of the latest signature and container signature
files.

This will now try to get the json from
nationalarchives.gov.uk/pronom/signatures.json This won't work at the
moment because that url doesn't exist but when we reverse proxy the old
site to the new page, it will and then we can quietly retire the SOAP
API. Anyone who wants to use it after that will need to upgrade.

I've added tests and updated the tests which were ignored so they work.